requirements.txt
alembic==0.9.2
beautifulsoup4==4.6.0
Chameleon==3.1
-e git+https://github.com/tianlantianlan95/CNX_Flip@22c74018ba54138fdf396202eed775c6631697d9#egg=cnx_flip
coverage==4.4.1
decorator==4.0.11
ez-setup==0.9
hupper==1.0
imageio==2.1.2
ipaddress==1.0.18
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
moviepy==0.2.3.2
music21==3.1.0
numpy==1.12.1
olefile==0.44
PasteDeploy==1.5.2
Pillow==4.1.1
psycopg2==2.7.1
py==1.4.33
Pygments==2.2.0
pyramid==1.8.3
pyramid-chameleon==0.3
pyramid-debugtoolbar==4.1
pyramid-jinja2==2.7
pyramid-mako==1.0.2
pyramid-tm==2.0
pytest==3.1.1
pytest-cov==2.5.1
python-dateutil==2.6.0
python-editor==1.0.3
repoze.lru==0.6
six==1.10.0
SQLAlchemy==1.1.10
tqdm==4.11.2
transaction==2.1.2
translationstring==1.3
unroll==0.1.0
venusian==1.1.0
waitress==1.0.2
WebOb==1.7.2
WebTest==2.0.27
zope.deprecation==4.2.0
zope.interface==4.4.1
zope.sqlalchemy==0.7.7
